+++ packages/core/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Dropped IE 11 support. Modern browsers now required.
 - Renamed the `annotation` field to `label`.
+- Renamed `fetchMetadata` to `fetchMessages`.
 
 #### 🚀 Updates