+++ packages/test-utils/CHANGELOG.md
@@ -15,6 +15,7 @@
 
 - Dropped Node.js v10 support. Requires v12.17 and above.
 - Renamed the `annotation` field to `label`.
+- Renamed `loadMeta` to `loadMessages`.
 
 #### 🚀 Updates